requirements.txt
requests==2.32.2
python-dotenv==0.19.1
numpy==1.26.4
pandas==1.5.3
openpyxl==3.0.9